rename network to local_network
