+++ zingo-cli/src/regtest.rs
@@ -4,7 +4,7 @@
 use std::path::PathBuf;
 use zingolib::testutils::scenarios::{LIGHTWALLETD_BIN, ZCASH_CLI_BIN, ZCASHD_BIN};
 use zingolib::testvectors::REG_O_ADDR_FROM_ABANDONART;
-use zingolib::wallet::network::ZingolibLocalNetwork;
+use zingolib::wallet::local_network::ZingolibLocalNetwork;
 use zingolib::zingo_infra_services::LocalNet;
 use zingolib::zingo_infra_services::indexer::{Lightwalletd, LightwalletdConfig};
 use zingolib::zingo_infra_services::validator::{Zcashd, ZcashdConfig};
