+++ zingolib/src/wallet.rs
@@ -25,7 +25,7 @@ use send::SendProgress;
 
 pub mod error;
 pub(crate) mod legacy;
-pub mod network;
+pub mod local_network;
 pub mod traits;
 pub mod utils;
