+++ zingolib/src/wallet/disk/testing/examples.rs
@@ -10,7 +10,7 @@ use super::super::LightWallet;
 use crate::config::ChainType;
 use crate::lightclient::LightClient;
 use crate::wallet::WalletSettings;
-use crate::wallet::network::ZingolibLocalNetwork;
+use crate::wallet::local_network::ZingolibLocalNetwork;
 
 /// ExampleWalletNetworkCase sorts first by Network, then seed, then last saved version.
 /// It is public so that any consumer can select and load any example wallet.
